+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/service/FilterSupportMaintainServiceImpl.java
@@ -5,7 +5,7 @@ import com.dwarfeng.fdr.stack.bean.entity.FilterSupport;
 import com.dwarfeng.fdr.stack.service.FilterSupportMaintainService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyEntireLookupService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyPresetLookupService;
-import com.dwarfeng.subgrade.impl.service.GeneralCrudService;
+import com.dwarfeng.subgrade.impl.service.GeneralBatchCrudService;
 import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionHelper;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
@@ -15,29 +15,42 @@ import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
 import com.dwarfeng.subgrade.stack.log.LogLevel;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Collections;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
 @Service
 public class FilterSupportMaintainServiceImpl implements FilterSupportMaintainService {
 
-    @Autowired
-    private GeneralCrudService<StringIdKey, FilterSupport> crudService;
-    @Autowired
-    private DaoOnlyEntireLookupService<FilterSupport> entireLookupService;
-    @Autowired
-    private DaoOnlyPresetLookupService<FilterSupport> presetLookupService;
+    private final GeneralBatchCrudService<StringIdKey, FilterSupport> crudService;
+    private final DaoOnlyEntireLookupService<FilterSupport> entireLookupService;
+    private final DaoOnlyPresetLookupService<FilterSupport> presetLookupService;
 
-    @SuppressWarnings("FieldMayBeFinal")
-    @Autowired(required = false)
-    private List<FilterSupporter> filterSupporters = Collections.emptyList();
+    private final List<FilterSupporter> filterSupporters;
 
-    @Autowired
-    private ServiceExceptionMapper sem;
+    private final ServiceExceptionMapper sem;
+
+    public FilterSupportMaintainServiceImpl(
+            GeneralBatchCrudService<StringIdKey, FilterSupport> crudService,
+            DaoOnlyEntireLookupService<FilterSupport> entireLookupService,
+            DaoOnlyPresetLookupService<FilterSupport> presetLookupService,
+            List<FilterSupporter> filterSupporters,
+            ServiceExceptionMapper sem
+    ) {
+        this.crudService = crudService;
+        this.entireLookupService = entireLookupService;
+        this.presetLookupService = presetLookupService;
+        if (Objects.isNull(filterSupporters)) {
+            this.filterSupporters = new ArrayList<>();
+        } else {
+            this.filterSupporters = filterSupporters;
+        }
+        this.sem = sem;
+    }
 
     @Override
     @BehaviorAnalyse
@@ -111,6 +124,88 @@ public class FilterSupportMaintainServiceImpl implements FilterSupportMaintainSe
 
     @Override
     @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(@SkipRecord List<StringIdKey> keys) throws ServiceException {
+        return crudService.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws ServiceException {
+        return crudService.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<FilterSupport> batchGet(@SkipRecord List<StringIdKey> keys) throws ServiceException {
+        return crudService.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public List<StringIdKey> batchInsert(@SkipRecord List<FilterSupport> elements) throws ServiceException {
+        return crudService.batchInsert(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchUpdate(@SkipRecord List<FilterSupport> elements) throws ServiceException {
+        crudService.batchUpdate(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(@SkipRecord List<StringIdKey> keys) throws ServiceException {
+        crudService.batchDelete(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<FilterSupport> batchGetIfExists(@SkipRecord List<StringIdKey> keys) throws ServiceException {
+        return crudService.batchGetIfExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public List<StringIdKey> batchInsertIfExists(@SkipRecord List<FilterSupport> elements) throws ServiceException {
+        return crudService.batchInsertIfExists(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchUpdateIfExists(@SkipRecord List<FilterSupport> elements) throws ServiceException {
+        crudService.batchUpdateIfExists(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDeleteIfExists(@SkipRecord List<StringIdKey> keys) throws ServiceException {
+        crudService.batchDeleteIfExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public List<StringIdKey> batchInsertOrUpdate(@SkipRecord List<FilterSupport> elements) throws ServiceException {
+        return crudService.batchInsertOrUpdate(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public PagedData<FilterSupport> lookup() throws ServiceException {
@@ -143,23 +238,22 @@ public class FilterSupportMaintainServiceImpl implements FilterSupportMaintainSe
 
     @Override
     @BehaviorAnalyse
-    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void reset() throws ServiceException {
-        for (FilterSupporter filterSupporter : filterSupporters) {
-            try {
-                crudService.insertOrUpdate(
-                        new FilterSupport(
-                                new StringIdKey(filterSupporter.provideType()),
-                                filterSupporter.provideLabel(),
-                                filterSupporter.provideDescription(),
-                                filterSupporter.provideExampleContent()
-                        )
-                );
-            } catch (Exception e) {
-                throw ServiceExceptionHelper.logAndThrow("重置触发器支持时发生异常",
-                        LogLevel.WARN, sem, e
-                );
-            }
+        try {
+            List<StringIdKey> filterKeys = entireLookupService.lookup().getData().stream()
+                    .map(FilterSupport::getKey).collect(Collectors.toList());
+            crudService.batchDelete(filterKeys);
+            List<FilterSupport> filterSupports = filterSupporters.stream().map(supporter -> new FilterSupport(
+                    new StringIdKey(supporter.provideType()),
+                    supporter.provideLabel(),
+                    supporter.provideDescription(),
+                    supporter.provideExampleContent()
+            )).collect(Collectors.toList());
+            crudService.batchInsert(filterSupports);
+        } catch (Exception e) {
+            throw ServiceExceptionHelper.logAndThrow(
+                    "重置过滤器支持时发生异常", LogLevel.WARN, sem, e
+            );
         }
     }
 }