+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/service/TriggerSupportMaintainServiceImpl.java
@@ -5,7 +5,7 @@ import com.dwarfeng.fdr.stack.bean.entity.TriggerSupport;
 import com.dwarfeng.fdr.stack.service.TriggerSupportMaintainService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyEntireLookupService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyPresetLookupService;
-import com.dwarfeng.subgrade.impl.service.GeneralCrudService;
+import com.dwarfeng.subgrade.impl.service.GeneralBatchCrudService;
 import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionHelper;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
@@ -15,29 +15,42 @@ import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
 import com.dwarfeng.subgrade.stack.log.LogLevel;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Collections;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
 @Service
 public class TriggerSupportMaintainServiceImpl implements TriggerSupportMaintainService {
 
-    @Autowired
-    private GeneralCrudService<StringIdKey, TriggerSupport> crudService;
-    @Autowired
-    private DaoOnlyEntireLookupService<TriggerSupport> entireLookupService;
-    @Autowired
-    private DaoOnlyPresetLookupService<TriggerSupport> presetLookupService;
+    private final GeneralBatchCrudService<StringIdKey, TriggerSupport> crudService;
+    private final DaoOnlyEntireLookupService<TriggerSupport> entireLookupService;
+    private final DaoOnlyPresetLookupService<TriggerSupport> presetLookupService;
 
-    @SuppressWarnings("FieldMayBeFinal")
-    @Autowired(required = false)
-    private List<TriggerSupporter> triggerSupporters = Collections.emptyList();
+    private final List<TriggerSupporter> triggerSupporters;
 
-    @Autowired
-    private ServiceExceptionMapper sem;
+    private final ServiceExceptionMapper sem;
+
+    public TriggerSupportMaintainServiceImpl(
+            GeneralBatchCrudService<StringIdKey, TriggerSupport> crudService,
+            DaoOnlyEntireLookupService<TriggerSupport> entireLookupService,
+            DaoOnlyPresetLookupService<TriggerSupport> presetLookupService,
+            List<TriggerSupporter> triggerSupporters,
+            ServiceExceptionMapper sem
+    ) {
+        this.crudService = crudService;
+        this.entireLookupService = entireLookupService;
+        this.presetLookupService = presetLookupService;
+        if (Objects.isNull(triggerSupporters)) {
+            this.triggerSupporters = new ArrayList<>();
+        } else {
+            this.triggerSupporters = triggerSupporters;
+        }
+        this.sem = sem;
+    }
 
     @Override
     @BehaviorAnalyse
@@ -111,6 +124,88 @@ public class TriggerSupportMaintainServiceImpl implements TriggerSupportMaintain
 
     @Override
     @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(@SkipRecord List<StringIdKey> keys) throws ServiceException {
+        return crudService.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws ServiceException {
+        return crudService.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<TriggerSupport> batchGet(@SkipRecord List<StringIdKey> keys) throws ServiceException {
+        return crudService.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public List<StringIdKey> batchInsert(@SkipRecord List<TriggerSupport> elements) throws ServiceException {
+        return crudService.batchInsert(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchUpdate(@SkipRecord List<TriggerSupport> elements) throws ServiceException {
+        crudService.batchUpdate(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(@SkipRecord List<StringIdKey> keys) throws ServiceException {
+        crudService.batchDelete(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<TriggerSupport> batchGetIfExists(@SkipRecord List<StringIdKey> keys) throws ServiceException {
+        return crudService.batchGetIfExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public List<StringIdKey> batchInsertIfExists(@SkipRecord List<TriggerSupport> elements) throws ServiceException {
+        return crudService.batchInsertIfExists(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchUpdateIfExists(@SkipRecord List<TriggerSupport> elements) throws ServiceException {
+        crudService.batchUpdateIfExists(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDeleteIfExists(@SkipRecord List<StringIdKey> keys) throws ServiceException {
+        crudService.batchDeleteIfExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public List<StringIdKey> batchInsertOrUpdate(@SkipRecord List<TriggerSupport> elements) throws ServiceException {
+        return crudService.batchInsertOrUpdate(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public PagedData<TriggerSupport> lookup() throws ServiceException {
@@ -143,23 +238,22 @@ public class TriggerSupportMaintainServiceImpl implements TriggerSupportMaintain
 
     @Override
     @BehaviorAnalyse
-    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void reset() throws ServiceException {
-        for (TriggerSupporter triggerSupporter : triggerSupporters) {
-            try {
-                crudService.insertOrUpdate(
-                        new TriggerSupport(
-                                new StringIdKey(triggerSupporter.provideType()),
-                                triggerSupporter.provideLabel(),
-                                triggerSupporter.provideDescription(),
-                                triggerSupporter.provideExampleContent()
-                        )
-                );
-            } catch (Exception e) {
-                throw ServiceExceptionHelper.logAndThrow("重置触发器支持时发生异常",
-                        LogLevel.WARN, sem, e
-                );
-            }
+        try {
+            List<StringIdKey> triggerKeys = entireLookupService.lookup().getData().stream()
+                    .map(TriggerSupport::getKey).collect(Collectors.toList());
+            crudService.batchDelete(triggerKeys);
+            List<TriggerSupport> triggerSupports = triggerSupporters.stream().map(supporter -> new TriggerSupport(
+                    new StringIdKey(supporter.provideType()),
+                    supporter.provideLabel(),
+                    supporter.provideDescription(),
+                    supporter.provideExampleContent()
+            )).collect(Collectors.toList());
+            crudService.batchInsert(triggerSupports);
+        } catch (Exception e) {
+            throw ServiceExceptionHelper.logAndThrow(
+                    "重置触发器支持时发生异常", LogLevel.WARN, sem, e
+            );
         }
     }
 }