+++ fdr-stack/src/main/java/com/dwarfeng/fdr/stack/service/MapperSupportMaintainService.java
@@ -3,7 +3,7 @@ package com.dwarfeng.fdr.stack.service;
 import com.dwarfeng.fdr.stack.bean.entity.MapperSupport;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
-import com.dwarfeng.subgrade.stack.service.CrudService;
+import com.dwarfeng.subgrade.stack.service.BatchCrudService;
 import com.dwarfeng.subgrade.stack.service.EntireLookupService;
 import com.dwarfeng.subgrade.stack.service.PresetLookupService;
 
@@ -13,7 +13,7 @@ import com.dwarfeng.subgrade.stack.service.PresetLookupService;
  * @author DwArFeng
  * @since 1.6.0
  */
-public interface MapperSupportMaintainService extends CrudService<StringIdKey, MapperSupport>,
+public interface MapperSupportMaintainService extends BatchCrudService<StringIdKey, MapperSupport>,
         EntireLookupService<MapperSupport>, PresetLookupService<MapperSupport> {
 
     String ID_LIKE = "id_like";